refactor(T1 E2d): move JSON post-build wiring (unit I) into USFExtendWiringService [BUILD-PASSED]
Third sub-commit of slice E2. Moves GenerateAndExecuteWiring (the ~1,650-line JSON wiring
orchestrator, incl. its local lambdas) + RegisterJsonBuiltActor/GetBuiltActorByCloneId/
GetSourceBuildableByName into SFExtendWiringService_Json.cpp (1,784 lines, <2k). SFExtendService.cpp
5,084 -> 3,397 (-1,687). 4 forwarders preserve the ~30 hologram/subsystem call sites + BP surface.
Friend ExtendService-> access to shared maps; nested USFExtendService::FSFSourcePoleWiringData
qualified for friend access. Shipping PackagePlugin clean (ExitCode=0) after fixing one nested-type
qualification. Static-verified: 0 unqualified members. Only unit F remains for slice E2.
